# Conflicts: #	pom.xml #	src/main/java/ru/modgy/booking/service/BookingServiceImpl.java #	src/main/java/ru/modgy/pet/controller/PetController.java #	src/main/java/ru/modgy/pet/service/PetServiceImpl.java #	src/main/java/ru/modgy/room/dto/mapper/RoomMapper.java #	src/main/java/ru/modgy/room/service/RoomServiceImpl.java #	src/test/java/ru/modgy/room/service/RoomServiceImplTest.java

+++ src/main/java/ru/modgy/pet/controller/PetController.java
@@ -12,6 +12,7 @@ import ru.modgy.pet.dto.NewPetDto;
 import ru.modgy.pet.dto.PetDto;
 import ru.modgy.pet.dto.UpdatePetDto;
 import ru.modgy.pet.service.PetService;
+import ru.modgy.utility.UtilityService;
 
 
 @CrossOrigin
@@ -20,34 +21,34 @@ import ru.modgy.pet.service.PetService;
 @RequiredArgsConstructor
 @Slf4j
 public class PetController {
-    private static final String USER_ID = "X-PetHotel-User-Id";
     private final PetService petService;
+    private final UtilityService utilityService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public PetDto addPet(
-            @RequestHeader(USER_ID) Long requesterId,
-            @RequestBody @Valid NewPetDto newPetDto) {
+    public PetDto addPet(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
+                         @RequestBody @Valid NewPetDto newPetDto) {
+        utilityService.checkBossAdminAccess(requesterId);
         log.info("PetController: POST/addPet, requesterId={}", requesterId);
         return petService.addPet(requesterId, newPetDto);
     }
 
     @GetMapping("/{id}")
-    public PetDto getPetById(
-            @RequestHeader(USER_ID) Long requesterId,
-            @PathVariable(value = "id") Long petId) {
+    public PetDto getPetById(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
+                             @PathVariable(value = "id") Long petId) {
         log.info("PetController: GET/getPetById, requesterId={}, petId={}", requesterId, petId);
         return petService.getPetById(requesterId, petId);
     }
 
     @GetMapping("/search")
     public Page<PetDto> getPetsBySearch(
-            @RequestHeader(USER_ID) Long requesterId,
+            @RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
             @RequestParam(required = false) String text,
             @PositiveOrZero
             @RequestParam(defaultValue = "0") Integer page,
             @Positive
             @RequestParam(defaultValue = "10") Integer size) {
+        utilityService.checkBossAdminAccess(requesterId);
         log.info("PetController: GET/getPetsBySearch, requesterId={}, text={}, page={}, size={}", requesterId,
                 text, page, size);
         return petService.getPetsBySearch(
@@ -58,19 +59,19 @@ public class PetController {
     }
 
     @PatchMapping("/{id}")
-    public PetDto updatePet(
-            @RequestHeader(USER_ID) Long requesterId,
-            @RequestBody @Valid UpdatePetDto updatePetDto,
-            @PathVariable(value = "id") Long petId) {
+    public PetDto updatePet(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
+                            @RequestBody @Valid UpdatePetDto updatePetDto,
+                            @PathVariable(value = "id") Long petId) {
+        utilityService.checkBossAdminAccess(requesterId);
         log.info("PetController: PATCH/updatePet, requesterId={}, petId={}", requesterId, petId);
         return petService.updatePet(requesterId, petId, updatePetDto);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deletePetById(
-            @RequestHeader(USER_ID) Long requesterId,
-            @PathVariable(value = "id") Long petId) {
+    public void deletePetById(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
+                              @PathVariable(value = "id") Long petId) {
+        utilityService.checkBossAdminAccess(requesterId);
         log.info("PetController: DELETE/deletePetById, requesterId= {}, petId={}", requesterId, petId);
         petService.deletePetById(requesterId, petId);
     }
